added CancellationTokenSource for serviceMessage

+++ GPTipsBot/Dtos/TelegramGptMessage.cs
@@ -27,7 +27,7 @@ namespace GPTipsBot.Dtos
         public string? Source { get; set; }
         public long? ContextId { get; internal set; }
         public DateTime CreatedAt { get; set; }
-        public int ServiceMessageId { get; set; }
+        public long ServiceMessageId { get; set; }
         public bool IsActive { get; set; }
         public bool ContextBound { get; set; }
     }
